For Colin Review

+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/console_pb/VersionedTextDocumentIdentifier.java
@@ -66,7 +66,7 @@ public class VersionedTextDocumentIdentifier {
 
   public native String getUri();
 
-  public native double getVersion();
+  public native int getVersion();
 
   public native Uint8Array serializeBinary();
